Run as a non-root user
Running as a non-root user is a security best practice. Some environments require containers run as non-root users.

For the `USER` directive, a numeric uid is specified instead of the username because systems configured to disallow running images as root aren't able to run images that use user name string values for the `USER` because they can't validate that a named user isn't root. See https://github.com/kubernetes/kubernetes/pull/56503 for details.